requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
langgraph>=0.2.50
langchain>=0.3
langchain-core>=0.3
langchain-openai>=0.2
langchain-mcp-adapters>=0.1.0
langgraph-checkpoint-redis>=0.0.5
langgraph-checkpoint-sqlite>=2.0
aiosqlite>=0.20
pyjwt>=2.8
httpx>=0.27
